requirements.txt
# general
flask

# models
Flask-SQLAlchemy
sqlalchemy-migrate

# assets
Flask-Assets
libsass
cssmin
jsmin

# miscellaneous
python-dotenv
CommonMark
Flask-WTF
Flask-Login
python-slugify

# documentation
eralchemy

# testing
mamba
expects
coverage
codeclimate-test-reporter
factory_boy

# utilities
pycodestyle
honcho
ipython

# database migrations
psycopg2

# cacheing
pylibmc
